+++ hot_fix_csx/lib/download_manager/resource_provider.dart
@@ -2,7 +2,7 @@
  * @Author: Cao Shixin
  * @Date: 2021-02-23 16:52:35
  * @LastEditors: Cao Shixin
- * @LastEditTime: 2022-11-10 18:11:00
+ * @LastEditTime: 2023-02-13 17:32:15
  * @Description: 网络资源处理工具
  * @Email: cao_shixin@yahoo.com
  * @Company: BrainCo
@@ -178,7 +178,7 @@ class ResourceProvider extends ChangeNotifier with SafeNotifier {
         _receivePort!.sendPort, 'downloader_send_port');
     _receivePort!.listen((dynamic data) {
       String id = data[0];
-      DownloadTaskStatus status = data[1];
+      DownloadTaskStatus status = DownloadTaskStatus(data[1] as int);
       int progress = data[2];
       var tempModel = _resourceMapLoadding[id];
       if (tempModel != null) {
@@ -200,7 +200,7 @@ class ResourceProvider extends ChangeNotifier with SafeNotifier {
               tableName: _tableName);
           _reloadArrStream();
         }
-        tempModel.refreshStreamControl.add(null);
+        tempModel.refreshStreamControl.sink.add(null);
       } else {
         FlutterDownloader.loadTasksWithRawQuery(
                 query: 'SELECT * FROM task WHERE task_id=$id')