+++ hot_fix_csx/lib/download_manager/resource_provider_tool.dart
@@ -2,7 +2,7 @@
  * @Author: Cao Shixin
  * @Date: 2021-03-18 15:03:02
  * @LastEditors: Cao Shixin
- * @LastEditTime: 2022-05-11 09:15:38
+ * @LastEditTime: 2023-02-13 17:32:03
  * @Description: 
  */
 import 'dart:async';
@@ -173,11 +173,12 @@ class ResourceProviderTool {
       }
     });
   }
-
+  
+  @pragma('vm:entry-point')
   static void downloadCallback(
       String id, DownloadTaskStatus status, int progress) {
     var send = IsolateNameServer.lookupPortByName('downloader_send_port');
-    send?.send([id, status, progress]);
+    send?.send([id, status.value, progress]);
   }
 }