+++ show_caseview_csx/lib/common/assets.dart
@@ -1,0 +1,14 @@
+/*
+ * @Author: Cao Shixin
+ * @Date: 2022-11-11 16:42:34
+ * @LastEditors: Cao Shixin
+ * @LastEditTime: 2022-11-11 16:46:45
+ * @Description: 
+ */
+const String _assets = 'assets/';
+
+class Images {
+  static const String prefix = '${_assets}images/';
+
+  static const String icCurve = '${prefix}ic_curve.svg';
+}