+++ show_caseview_csx/lib/common/const.dart
@@ -1,0 +1,9 @@
+/*
+ * @Author: Cao Shixin
+ * @Date: 2022-11-11 16:47:08
+ * @LastEditors: Cao Shixin
+ * @LastEditTime: 2022-11-11 16:47:09
+ * @Description: 
+ */
+
+const packageName = 'show_caseview_csx';