+++ hot_fix_csx/CHANGELOG.md
@@ -2,13 +2,18 @@
  * @Author: Cao Shixin
  * @Date: 2022-04-20 09:01:23
  * @LastEditors: Cao Shixin
- * @LastEditTime: 2022-11-10 18:13:55
+ * @LastEditTime: 2023-02-13 17:36:04
  * @Description: 
 -->
-## 0.0.8+2 20221105
+## 0.0.9 20230213
+* 适配flutter3.7版本导致flutter_downloader下载状态无法回调问题
+
+
+## 0.0.8+4 20221105
 * 修复批量下载中有暂停资源问题
 * 升级flutter 3.*版本
 * 添加资源下载oom检测以及toast提醒功能
+* 添加flutter_downloader注解，避免release环境下载状态不回调
 
 
 ## 0.0.7 20220819