+++ show_caseview_csx/CHANGELOG.md
@@ -2,9 +2,15 @@
  * @Author: Cao Shixin
  * @Date: 2022-11-11 14:18:13
  * @LastEditors: Cao Shixin
- * @LastEditTime: 2022-11-11 14:54:27
+ * @LastEditTime: 2022-11-12 17:22:04
  * @Description: 
 -->
+
+## 0.0.2 20221112
+* 扩展指向小三角支持曲线+箭头
+* 扩展target处在蒙层之上的自定义组件扩展
+* tooltip的强制在target之上或之下，支持偏移offset扩展
+
 ## 0.0.1 20221111
 
 * 迁移from showcaseview 2.0.0+1 源地址：https://github.com/simformsolutions/flutter_showcaseview